Basic Client
Implement the basics for supporting the Couch HTTP interface with
primitives for Server, Database and Document objects. Define Servers as
being a collection of Databases and Databases as being a collection of
Documents. Define simple Resource class as a lightweight wrapper around
the requests library filling in a few defaults for common headers Couch
expects to see during requests.